Time: 34 ms (48.86%), Space: 94.6 MB (80.97%) - LeetHub

+++ 1838-frequency-of-the-most-frequent-element/1838-frequency-of-the-most-frequent-element.java
@@ -5,7 +5,7 @@ class Solution {
         Arrays.sort(nums);
         for (int i = 0; i < nums.length; i++) {
             sum += nums[i];
-            long x = nums[i] * (i - j + 1);
+            long x = (long)nums[i] * (i - j + 1);
             while (x - sum > k) {
                 sum -= nums[j++];
                 x -=nums[i];
